Drastic improvement (and a total bc break with any existing draft content, sorry - hey folks this plugin is totally unannounced at this stage):
This plugin is now implemented using a separate draft_version column in the a_area table, complementing the latest_version column. latest_version is used for all logged-out users and in "applied" mode. The greater of draft_version and latest_version (or whichever is not null) is used in "draft" mode. This is a substantial improvement over the first implementation of this plugin, which used separate virtual pages as stand-ins for each page, because it requires far less code and the upgrade is painless, with no need for a task to sync "back" to draft.



git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4707 c89fab08-2f9b-42bf-8176-bd3ad6bee547