The new functionality for overriding a template via a partial or component had a big problem: there was no way to decorate the template or partial with the layout.
I fixed this by doing the include of the partial or component from an actual template. This ensures that the layout is applied normally. Note that it doesn't stop you from shutting off the layout for a particular action as you normally would.

It also means that in modules in which you choose to take advantage of this new functionality you must have the renderTemplateViaComponent and renderTemplateViaPartial templates as found in the aMedia module.

(There is no requirement to use this new functionality, it never comes into play for modules that don't extend aEngineActions, and even if they do, it only comes into play if you configure overrides in app.yml. So backwards compatibility is not disturbed.)




git-svn-id: http://svn.apostrophenow.org/plugins/apostrophePlugin/branches/1.5@4749 c89fab08-2f9b-42bf-8176-bd3ad6bee547